BT073 - feat: get all categories
Moving method getAllCategories() from CategoryController to AdminController.
