+++ src/main/java/pl/envelo/meetek/controller/user/AdminController.java
@@ -12,8 +12,11 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import pl.envelo.meetek.dto.CategoryDto;
 import pl.envelo.meetek.dto.event.SingleEventShortDto;
+import pl.envelo.meetek.model.Category;
 import pl.envelo.meetek.model.event.SingleEvent;
+import pl.envelo.meetek.service.CategoryService;
 import pl.envelo.meetek.service.DtoMapperService;
 import pl.envelo.meetek.service.event.SingleEventService;
 
@@ -22,10 +25,12 @@ import java.util.stream.Collectors;
 
 @AllArgsConstructor
 @RestController
-@Tag(name = "User")
+@Tag(name = "Admin")
 @RequestMapping("/${app.prefix}/${app.version}/admin")
 public class AdminController {
+
     private SingleEventService singleEventService;
+    private final CategoryService categoryService;
     private DtoMapperService dtoMapperService;
 
     @GetMapping("/past-events")
@@ -75,6 +80,21 @@ public class AdminController {
         } else {
             return new ResponseEntity(futureEventsShortDto, HttpStatus.OK);
         }
+    }
+
+    @GetMapping("/categories")
+    @Operation(summary = "Get all categories")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Results returned",
+                    content = {@Content(mediaType = "application/json", schema = @Schema(implementation = CategoryDto.class))}),
+            @ApiResponse(responseCode = "204", description = "No category found", content = @Content)})
+    public ResponseEntity<List<CategoryDto>> getAllCategories() {
+        List<Category> categories = categoryService.getAllCategories();
+        List<CategoryDto> dtoCategories = categories.stream()
+                .map(dtoMapperService::mapToCategoryDto)
+                .toList();
+        HttpStatus status = dtoCategories.isEmpty() ? HttpStatus.NO_CONTENT : HttpStatus.OK;
+        return new ResponseEntity<>(dtoCategories, status);
     }
 
 }
